made the select image component reuseable, between gallery and staff component
