+++ src/app/home/component/ourStaff.tsx
@@ -1,11 +1,17 @@
+"use client";
 import ComponentsHeaders from "~/app/shared/components/componentsHeaders";
 import quoteIcon from "public/quote-icon.png";
 import Image from "next/image";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { staffData, staffDataSchema } from "@/shared/data/staffData";
 import { faFacebook } from "@fortawesome/free-brands-svg-icons";
+import SelectedImage from "../../shared/components/selectedImage";
+import { useState } from "react";
 
 export default function OurStaff() {
+  const [showImage, setShowImage] = useState(false);
+  const [imageData, setImageData] = useState({});
+
   return (
     <div className="flex h-auto flex-col items-center pb-[50px] pt-[60px] text-center">
       <ComponentsHeaders
@@ -28,9 +34,15 @@ export default function OurStaff() {
                 key={index}
                 className="flex items-center justify-center gap-[10px] text-start transition-all  duration-500 phone:h-[600px] phone:w-[360px] phone:flex-col laptop:h-[333px] laptop:w-[509px] laptop:flex-row"
               >
-                <div className="relative flex h-[100%] flex-col text-center transition-all duration-500 hover:-translate-y-3 hover:bg-[#f1f1f1af] phone:w-[100%] laptop:w-[50%] ">
+                <div className=" relative flex h-[100%] flex-col text-center transition-all duration-500 hover:-translate-y-3 hover:bg-[#f1f1f1af] phone:w-[100%] laptop:w-[50%] ">
                   <div className="relative h-[70%]">
-                    <div className="absolute left-0 top-0 z-[10] flex h-[100%] w-[100%] flex-col items-center justify-center text-center opacity-0 transition-all duration-500 hover:opacity-100 ">
+                    <div
+                      onClick={() => {
+                        setShowImage(true);
+                        setImageData(data);
+                      }}
+                      className=" absolute  left-0 top-0 z-[10] flex h-[100%] w-[100%] cursor-pointer flex-col items-center justify-center text-center opacity-0 transition-all duration-500 hover:opacity-100 "
+                    >
                       <div className=" flex flex-col items-center justify-center rounded-[50%] bg-[#57509f4f] text-center phone:h-[90px] phone:w-[90px] laptop:h-[150px] laptop:w-[150px] ">
                         <Image
                           src={data.galleryIconPlus}
@@ -101,6 +113,14 @@ export default function OurStaff() {
             return null;
           }
         })}
+      </div>
+      <div className={`${!showImage && "hidden"} `}>
+        <SelectedImage
+          data={imageData}
+          showImage={setShowImage}
+          galleryData={""}
+          setImageData={""}
+        />
       </div>
     </div>
   );
