+++ src/app/shared/data/staffData.ts
@@ -35,7 +35,7 @@ export const staffData = [
     role: "Assistant Teacher",
     description:
       "My name is Katie. I grew up and studied in Canada. This is my second year at Fable and love every minute of it, making the children’s learning experience fun.",
-    id: 1,
+    id: 2,
     socialMedia: [
       { icon: faFacebook, url: "/" },
       { icon: faInstagram, url: "/" },
@@ -49,7 +49,7 @@ export const staffData = [
     role: "Lead Teacher",
     description:
       "I hold a degree in Early Childhood Education and an advanced English language certificate. I have been working as a kindergarten teacher since 2002.",
-    id: 1,
+    id: 3,
     socialMedia: [
       { icon: faFacebook, url: "/" },
       { icon: faInstagram, url: "/" },
@@ -63,7 +63,7 @@ export const staffData = [
     role: "Teacher",
     description:
       "I have completed a Graduate Diploma in Early Childhood Teaching. I have worked with children aged from six weeks to eight years. This is my second year at Fable.",
-    id: 1,
+    id: 4,
     socialMedia: [
       { icon: faFacebook, url: "/" },
       { icon: faInstagram, url: "/" },
